# lib/api-zod/src/generated/types/missionary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MissionaryLatestPostType } from './missionaryLatestPostType';

export interface Missionary {
  id: string;
  userId: string;
  name: string;
  /** Omitted when the missionary hides their email address. */
  email?: string;
  /** Omitted when the missionary hides their biography. */
  bio?: string;
  /** Omitted when the missionary hides their location. */
  country?: string;
  initials: string;
  isFollowed: boolean;
  /** @nullable */
  latestPostType?: MissionaryLatestPostType;
}